source code folder
Theme: Electronic Voting System
Problem Statement: Voting is a method for a group of individuals to make a collective decision or voice their cumulative opinion to arrive at a consensus. The results of a vote can have far reaching consequences due to which it is imperative to maintain Integrity by ensuring that there is no scope for fraud or cheating to occur while the votes are being cast. It is highly crucial that voting is carried out in a fair and just manner.
Goal: 
•	User cannot know any information about the vote
•	Votes cannot be tampered